Moving dayno function from aacgm_v2 library to general rtime library
