--- codebase/analysis/src.lib/aacgm_v2/aacgm.1.0/include/genmag.h
@@ -1,2 +1,0 @@
-int dayno(int year, int month, int day, int *diy);
-
