+++ codebase/general/src.lib/time.1.7/include/rtime.h
@@ -31,20 +31,21 @@ Modifications:
 
 void TimeYrsecToYMDHMS(int time,int yr, 
                         int *mo,int *dy,
-			int *hr,int *mn,int *sc);
+                        int *hr,int *mn,int *sc);
 int TimeYMDHMSToYrsec(int yr, int mo, int day, int hr, int mn,int sec);
 double TimeYMDHMSToEpoch(int yr,int mo,int dy,int hr,int mn,double sec);
 void TimeEpochToYMDHMS(double tme,int *yr,int *mo,int *dy,int *hr,int *mn,
-		       double *sc);
+                       double *sc);
 double TimeYMDHMSToJulian(int yr,int mo,int dy,int hr,int mt,double sc);
 int TimeJulianToYMDHMS(double jd,int *yr,int *mo,
-		       int *dy,int *hr,int *mt,double *sc);
+                       int *dy,int *hr,int *mt,double *sc);
 
 void TimeReadClock(int *yr,int *month,int *day,int *hour,int *min,int *sec,
-		   int *usec);
+                   int *usec);
 
 double TimeStrToSOD(char *text);
 double TimeStrToEpoch(char *text);
+int dayno(int year, int month, int day, int *diy);
 
 
 #endif
